Backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=1.4
alembic
psycopg2-binary
python-dotenv
# passlib[bcrypt]
python-jose[cryptography]
pydantic
email-validator
python-multipart
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
alembic